4f is now bonus Q, changed it to make easier
Previously, the region was chr8:128,740,000-128,760,000 but this is a larger region and made the problem much harder. The new region is chr8:129127000-129128000. 
This question is now just for bonus marks. For manually reporting the haplotypes, you will receive 4 bonus points. If you provide code to generate the haplotype sequences, then you get an additional 2 bonus points. 

Successfully answering 4f can give you up to 6 bonus points but you may not score over 100% on the assignment.